requirements.txt
streamlit==1.35.0
rapidfuzz==3.9.3
google-api-python-client==2.131.0
google-auth==2.29.0
pandas>=1.3.0